Changed Doubles to Float. Fixed API-nutrients

+++ src/TestingSignup.java
@@ -83,10 +83,10 @@ class FakeDAO implements SignupDataAccessInterface, RecommendDataAccessInterface
     }
 
     @Override
-    public HashMap<String, Double> getFoodNutritionalValues(String foodName, Float quantity) {
-        HashMap<String, Double> nutritionalValues = new HashMap<String, Double>();
-        nutritionalValues.put("Carbs", 10.0);
-        nutritionalValues.put("Protein", 5.0);
+    public HashMap<String, Float> getFoodNutritionalValues(String foodName, Float quantity) {
+        HashMap<String, Float> nutritionalValues = new HashMap<String, Float>();
+        nutritionalValues.put("Carbs", 10.0f);
+        nutritionalValues.put("Protein", 5.0f);
         return nutritionalValues;
     }
 
@@ -100,7 +100,7 @@ class FakeDAO implements SignupDataAccessInterface, RecommendDataAccessInterface
 class FakeUserFactory implements entity.UserFactory{
 
     @Override
-    public User create(String username, String password, String name, String gender, Double weight, int age, Double height, ArrayList<String> dietaryRestrictions) {
+    public User create(String username, String password, String name, String gender, Float weight, int age, Float height, ArrayList<String> dietaryRestrictions) {
         BasicUser basicUser = new BasicUser(username, password, name, gender, weight, age, height, dietaryRestrictions);
         System.out.println("made user");
         return basicUser;
